requirements.txt
# ─── Core ────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.1
requests>=2.32.3
PyYAML>=6.0.2
schedule>=1.2.2

# ─── ML Classification ────────────────────────────────────────────────────────
scikit-learn>=1.6.1
numpy>=1.24.1
joblib>=1.3.0

# ─── PostgreSQL Audit Logging ─────────────────────────────────────────────────
psycopg2-binary>=2.9.9

# ─── Dev / Testing ────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-mock>=3.14.0
